# backend/src/domains/auth/auth.routes.js
import { Router } from "express";
import { success } from "../../shared/http/response.js";
import {
  adminDevLogin,
  refreshAccessToken,
  sendPartnerOtp,
  sendUserOtp,
  verifyPartnerOtp,
  verifyUserOtp,
  logoutSession,
} from "./auth.service.js";
import {
  adminDevLoginSchema,
  refreshTokenSchema,
  sendOtpSchema,
  verifyOtpSchema,
  verifyUserOtpSchema,
  logoutSchema,
} from "./auth.schemas.js";

export const authRouter = Router();

authRouter.post("/partner/otp/send", (req, res, next) => {
  try {
    const input = sendOtpSchema.parse(req.body);
    const result = sendPartnerOtp(input.phone);
    res.json(success(result));
  } catch (err) {
    next(err);
  }
});

authRouter.post("/partner/otp/verify", (req, res, next) => {
  try {
    const input = verifyOtpSchema.parse(req.body);
    const result = verifyPartnerOtp(input);
    res.json(success(result));
  } catch (err) {
    next(err);
  }
});

authRouter.post("/refresh", (req, res, next) => {
  try {
    const input = refreshTokenSchema.parse(req.body);
    const result = refreshAccessToken(input.refreshToken);
    res.json(success(result));
  } catch (err) {
    next(err);
  }
});

authRouter.post("/logout", (req, res, next) => {
  try {
    const input = logoutSchema.parse(req.body);
    const result = logoutSession(input.refreshToken);
    res.json(success(result));
  } catch (err) {
    next(err);
  }
});

authRouter.post("/user/otp/send", (req, res, next) => {
  try {
    const input = sendOtpSchema.parse(req.body);
    const result = sendUserOtp(input.phone);
    res.json(success(result));
  } catch (err) {
    next(err);
  }
});

authRouter.post("/user/otp/verify", (req, res, next) => {
  try {
    const input = verifyUserOtpSchema.parse(req.body);
    const result = verifyUserOtp(input);
    res.json(success(result));
  } catch (err) {
    next(err);
  }
});

// Dev-only helper so admin serviceability APIs can be tested from day one.
authRouter.post("/admin/dev-login", (req, res, next) => {
  try {
    const input = adminDevLoginSchema.parse(req.body);
    const result = adminDevLogin(input);
    res.json(success(result));
  } catch (err) {
    next(err);
  }
});